+++ sdk/mosaic_model.go
@@ -26,3 +26,30 @@ const (
 	DECREASE MosaicSupplyType = 0
 	INCREASE MosaicSupplyType = 1
 )
+
+type NamespaceMosaicMetaDTO struct {
+	Active bool
+	Index  int
+	Id     string
+} /* NamespaceMosaicMetaDTO */
+func (ref *NamespaceInfoDTO) extractLevels() []*NamespaceId {
+
+	levels := make([]*NamespaceId, 3)
+	var err error
+
+	nsId := NewNamespaceId(ref.Namespace.Level0, "")
+	if err == nil {
+		levels = append(levels, nsId)
+	}
+
+	nsId = NewNamespaceId(ref.Namespace.Level1, "")
+	if err == nil {
+		levels = append(levels, nsId)
+	}
+
+	nsId = NewNamespaceId(ref.Namespace.Level2, "")
+	if err == nil {
+		levels = append(levels, nsId)
+	}
+	return levels
+}
